fix : video publish
