+++ TIKTOK_Video/api/handler/videoHandler.go
@@ -9,6 +9,7 @@ import (
 	"context"
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/cloudwego/hertz/pkg/protocol/consts"
+	"log"
 	"net/http"
 	"strconv"
 	"time"
@@ -73,6 +74,30 @@ func Feed(ctx context.Context, c *app.RequestContext) {
 	})
 }
 
+// PublishAction
 /*
-	评论操作接口
+	登录用户选择视频上传
 */
+func PublishAction(ctx context.Context, c *app.RequestContext) {
+	// get the basic info from meta
+	//user, _ := c.Get(mw.IdentityKey)
+	videoTitle := c.PostForm("title")
+	videoData, err := c.Request.FormFile("data")
+	if err != nil {
+		log.Print("can not get this filestream")
+	}
+	vsi := ServiceImpl.VideoServiceImpl{}
+	if err := vsi.PublishVideo(24, videoData, videoTitle); err != nil {
+		c.JSON(consts.StatusOK, vo.Response{
+			StatusCode: ResponseFail,
+			StatusMsg:  err.Error(),
+		})
+		return
+	}
+
+	c.JSON(consts.StatusOK, vo.Response{
+		StatusCode: ResponseSuccess,
+		StatusMsg:  "publish success!",
+	})
+
+}
