+++ TIKTOK_Video/service/ServiceImpl/videoServiceImpl.go
@@ -2,12 +2,13 @@ package ServiceImpl
 
 import (
 	"TIKTOK_Video/dal/mysql"
+	"TIKTOK_Video/model"
 	"TIKTOK_Video/model/vo"
 	"TIKTOK_Video/mw"
 	"bytes"
 	"errors"
 	"fmt"
-	"github.com/satori/go.uuid"
+	uuid "github.com/satori/go.uuid"
 	ffmpeg "github.com/u2takey/ffmpeg-go"
 	"image"
 	"image/jpeg"
@@ -15,12 +16,75 @@ import (
 	"log"
 	"mime/multipart"
 	"os"
+	"time"
 )
 
-type PublishServiceImpl struct {
+type VideoServiceImpl struct {
 }
 
-func (psi *PublishServiceImpl) PublishVideo(userId int64, fileHeader *multipart.FileHeader, videoTitle string) error {
+func (vsi *VideoServiceImpl) GetVideoInfosByLatestTime(latestTime int64, userName string) ([]vo.VideoInfo, int64, error) {
+	var videoInfos []vo.VideoInfo
+	nextTime := time.Now().Unix()
+
+	videos, err := mysql.GetVideosByTime(latestTime)
+	if err != nil {
+		return videoInfos, nextTime, err
+	}
+
+	if len(videos) == 0 {
+		return videoInfos, nextTime, errors.New("video is empty")
+	}
+	videoInfos = make([]vo.VideoInfo, len(videos))
+	videoInfos, err = bindVideoInfo(videoInfos, videos, userName)
+
+	nextTime = videos[len(videos)-1].PublishTime
+	return videoInfos, nextTime, err
+}
+
+func bindVideoInfo(videoInfos []vo.VideoInfo, videos []model.Video, userName string) ([]vo.VideoInfo, error) {
+	for i, video := range videos {
+		var err error
+		videoId := video.VideoId
+
+		favoriteCount, err := mysql.GetFavoriteCountByID(videoId)
+		if err != nil {
+			return videoInfos, err
+		}
+		commentCount, err := mysql.GetCommentCountByID(videoId)
+		if err != nil {
+			return videoInfos, err
+		}
+		// 需要与user通信，应定义到service层
+		var favorite bool
+		if userName != "" {
+			favorite, err = isFavorite(videoId, userName)
+			if err != nil {
+				return videoInfos, err
+			}
+		}
+
+		videoInfos[i].Id = videoId
+		videoInfos[i].Author = vo.DemoUser
+		//videoInfos[i].Author.Id = video.AuthorId
+		videoInfos[i].PlayUrl = video.PlayUrl
+		videoInfos[i].CoverUrl = video.CoverUrl
+		videoInfos[i].FavoriteCount = favoriteCount
+		videoInfos[i].CommentCount = commentCount
+		videoInfos[i].IsFavorite = favorite
+		videoInfos[i].Title = video.Title
+
+	}
+
+	return videoInfos, nil
+}
+
+// 调用远程接口，判断userName是否喜欢videoId视频
+func isFavorite(videoId int64, userName string) (bool, error) {
+	// TODO : impl
+	return false, nil
+}
+
+func (vsi *VideoServiceImpl) PublishVideo(userId int64, fileHeader *multipart.FileHeader, videoTitle string) error {
 	// 处理multipart.FileHeader文件为byte[]
 	file, err := fileHeader.Open()
 	buf := bytes.NewBuffer(nil)
@@ -77,6 +141,7 @@ func CreateVideoCall(userId int64, videoData []byte, videoTitle string) (int64, 
 	}
 	return videoId, nil
 }
+
 func readFrameAsJpeg(filePath string) ([]byte, error) {
 	reader := bytes.NewBuffer(nil)
 	//根据对应的url切帧
@@ -99,41 +164,6 @@ func readFrameAsJpeg(filePath string) ([]byte, error) {
 	}
 	//返回图像的byte[]流
 	return buf.Bytes(), err
-}
-
-/*
-GetVideoList
-获取userId发布的视频列表
-
-@ params
-
-	userIdTar: 请求查询的userId
-	userIdSrc: 登录userId
-*/
-func (psi *PublishServiceImpl) GetVideoList(userIdTar int64, userIdSrc int64) ([]vo.VideoInfo, error) {
-	// 调用Dao层，查找userIdTar用户的所有videoIds
-	var videoInfos []vo.VideoInfo
-	videoIds, err := mysql.GetPublishVideoIdsById(userIdTar)
-	if err != nil {
-		return videoInfos, err
-	}
-	if len(videoIds) == 0 {
-		return videoInfos, errors.New("VideoList is empty")
-	}
-	videoInfos, err = getVideoInfosByVideoIds(videoIds, userIdSrc, "publish_query")
-	return videoInfos, err
-}
-
-/*
-通过videoIds封装VideoInfo，发布视频列表和喜欢列表可用
-@ params
-
-	videoIds: 视频ids
-	userId: 登录的userId，用于查询登录账号与<视频(是否喜欢)和用户(是否关注)>的关系
-*/
-func getVideoInfosByVideoIds(videoIds []int64, userId int64, mode string) ([]vo.VideoInfo, error) {
-
-	return nil, nil
 }
 
 /*
