--- TIKTOK_Video/service/publishService.go
@@ -1,5 +1,0 @@
-package service
-
-type PublishService interface {
-	PublishVideo(userId int64, videoData []byte, videoTitle string) error
-}
